pyproject.toml
[project]
name = "mewgent"
version = "0.1.0"
description = "Live Mewgenics companion app — OCR cat stats from screen capture"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6",
    "opencv-contrib-python>=4.9",
    "easyocr>=1.7",
    "imagehash>=4.3",
    "pywin32>=306",
    "pyyaml>=6.0",
    "Pillow>=10.0",
    "numpy>=1.26",
    "torch>=2.2",
    "torchvision>=0.17",
]

[tool.uv]
override-dependencies = [
    "opencv-python-headless; python_version < '0'",
    "opencv-python; python_version < '0'",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cu121" }]
torchvision = [{ index = "pytorch-cu121" }]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true
